+++ R/dynSurv.R
@@ -23,25 +23,65 @@
 #'   measurement time. If omitted (default is \code{u=NULL}), then conditional
 #'   failure probabilities are reported for all observed failure times in the
 #'   \code{mjoint} object data.
+#' @param type a character string for whether a first-order
+#'   (\code{type="first-order"}) or Monte Carlo simulation approach
+#'   (\code{type="simulated"}) should be used for the dynamic prediction.
+#'   Defaults to the computationally faster first-order prediction method.
+#' @param M for \code{type="simulated"}, the number of simulations to performs.
+#'   Default is \code{M=200}.
+#' @param scale a numeric scalar that scales the variance parameter of the
+#'   proposal distribution for the Metropolis-Hastings algorithm, which
+#'   therefore controls the acceptance rate of the sampling algorithm.
+#' @param ci a numeric value with value in the interval \eqn{(0, 1)} specifying
+#'   the confidence interval level for predictions of \code{type='simulated'}.
+#'   If missing, defaults to \code{ci=0.95} for a 95\% confidence interval. If
+#'   \code{type='first-order'} is used, then this argument is ignored.
+#' @param progress logical: should a progress bar be shown on the console to
+#'   indicate the percentage of simulations completed? Default is
+#'   \code{progress=TRUE}.
 #'
 #' @details Dynamic predictions for the time-to-event data sub-model based on an
 #'   observed measurement history for the longitudinal outcomes of a new subject
-#'   are based on a first-order approximation described in Rizopoulos (2011).
-#'   Namely, given that the subject was last observed at time \emph{t}, we
-#'   calculate the conditional survival probability at time \eqn{u > t} as
+#'   are based on either a first-order approximation or Monte Carlo simulation
+#'   approach, both of which are described in Rizopoulos (2011). Namely, given
+#'   that the subject was last observed at time \emph{t}, we calculate the
+#'   conditional survival probability at time \eqn{u > t} as
 #'
-#'   \deqn{P[T \ge u | T \ge t; y, \theta] \approx
-#'   \frac{S(u | \hat{b}; \theta)}{S(t | \hat{b}; \theta)},}
+#'   \deqn{P[T \ge u | T \ge t; y, \theta] \approx \frac{S(u | \hat{b};
+#'   \theta)}{S(t | \hat{b}; \theta)},}
 #'
 #'   where \eqn{T} is the failure time for the new subject, \eqn{y} is the
-#'   stacked-vector of longitudinal measurements up to time \emph{t}, \eqn{S(u |
-#'   \hat{b}; \theta)} is the survival function, and \eqn{\hat{b}} is the mode
+#'   stacked-vector of longitudinal measurements up to time \emph{t} and
+#'   \eqn{S(u | \hat{b}; \theta)} is the survival function.
+#'
+#'   \strong{First order predictions}
+#'
+#'   For \code{type="first-order"}, \eqn{\hat{b}} is the mode
 #'   of the posterior distribution of the random effects given by
 #'
 #'   \deqn{\hat{b} = {\arg \max}_b f(b | y, T \ge t; \theta).}
 #'
 #'   The predictions are based on plugging in \eqn{\theta = \hat{\theta}}, which
 #'   is extracted from the \code{mjoint} object.
+#'
+#'   \strong{Monte Carlo simulation predictions}
+#'
+#'   For \code{type="simulated"}, \eqn{\theta} is drawn from a multivariate
+#'   normal distribution with means \eqn{\hat{\theta}} and variance-covariance
+#'   matrix both extracted from the fitted \code{mjoint} object via the
+#'   \code{coef()} and \code{vcov()} functions. \eqn{\hat{b}} is drawn from the
+#'   the posterior distribution of the random effects
+#'
+#'   \deqn{f(b | y, T \ge t; \theta)}
+#'
+#'   by means of a Metropolis-Hasting algorithm with independent multivariate
+#'   non-central \emph{t}-distribution proposal distributions with
+#'   non-centrality parameter \eqn{\hat{b}} from the first-order prediction and
+#'   variance-covariance matrix equal to \code{scale} \eqn{\times} the inverse
+#'   of the negative Hessian of the posterior distribution. The choice os
+#'   \code{scale} can be used to tune the acceptance rate of the
+#'   Metropolis-Hastings sampler. This simulation algorithm is iterated \code{M}
+#'   times, at each time calculating the conditional survival probability.
 #'
 #' @author Graeme L. Hickey (\email{graeme.hickey@@liverpool.ac.uk})
 #' @keywords survival
@@ -59,9 +99,13 @@
 #' models. \emph{Biometrics}. 2013; \strong{69}: 206–13.
 #'
 #' @return A list object inheriting from class \code{dynSurv}. The list returns
-#'   the arguments of the function and a \code{data.frame} of 2 columns, with
-#'   first column (named \code{u}) denoting times and the second column (named
-#'   \code{surv}) denoting the conditional failure probability.
+#'   the arguments of the function and a \code{data.frame} with first column
+#'   (named \code{u}) denoting times and the subsequent columns returning
+#'   summary statistics for the conditional failure probabilities For
+#'   \code{type="first-order"}, a single column named \code{surv} is appended.
+#'   For \code{type="simulated"}, four columns named \code{mean}, \code{median},
+#'   \code{lower} and \code{upper} are appended, denoting the mean, median and
+#'   lower and upper confidence intervals from the Monte Carlo draws.
 #' @export
 #'
 #' @examples
@@ -85,11 +129,17 @@
 #' hvd2 <- droplevels(hvd[hvd$num == 1, ])
 #' dynSurv(fit2, hvd2)
 #' dynSurv(fit2, hvd2, u = 7) # survival at 7-years only
+#' dynSurv(fit2, hvd2, type = "simulate", M = 2)
 #' }
-dynSurv <- function(object, newdata, newSurvData = NULL, u = NULL) {
+dynSurv <- function(object, newdata, newSurvData = NULL, u = NULL,
+                    type = "first-order", M = 200, scale = 2, ci, progress = TRUE) {
 
   if (!inherits(object, "mjoint")) {
     stop("Use only with 'mjoint' model objects.\n")
+  }
+
+  if (type != "first-order" && type != "simulated") {
+    stop("order must be specified as either 'first-order' or 'simulated")
   }
 
   data.t <- process_newdata(newdata = newdata,
@@ -111,36 +161,130 @@ dynSurv <- function(object, newdata, newSurvData = NULL, u = NULL) {
   }
 
   # Get posterior mode of [b | data, theta]
-  b.hat <- b.mode(data = data.t, theta = object$coefficients)
+  b.hat <- b_mode(data = data.t, theta = object$coefficients)
 
-  if (b.hat$convergence == 0) {
-    b.hat <- b.hat$par
-  } else {
+  if (b.hat$convergence != 0) {
     stop("Optimisation failed")
   }
 
-  # Predicted survival at time t
-  S.t <- S(b = b.hat, data = data.t, theta = object$coefficients)
+  #***********************************************
+  # First-order prediction
+  #***********************************************
 
-  # Predicted survival at times u
-  S.u <- vector(length = length(u))
-  for (i in 1:length(u)) {
-    data.u <- process_newdata(newdata = newdata,
-                              object = object,
-                              tobs = u[i],
-                              newSurvData = newSurvData)
+  if (type == "first-order") {
 
-    S.u[i] <- S(b = b.hat, data = data.u, theta = object$coefficients)
-  }
+    b.hat <- b.hat$par
 
-  pred <- data.frame("u" = u, "surv" = S.u / S.t)
+    # Predicted survival at time t
+    S.t <- S(b = b.hat, data = data.t, theta = object$coefficients)
+
+    # Predicted survival at times u
+    S.u <- vector(length = length(u))
+    for (i in 1:length(u)) {
+      data.u <- process_newdata(newdata = newdata,
+                                object = object,
+                                tobs = u[i],
+                                newSurvData = newSurvData)
+
+      S.u[i] <- S(b = b.hat, data = data.u, theta = object$coefficients)
+    }
+
+    pred <- data.frame("u" = u, "surv" = S.u / S.t)
+
+  } # end first-order prediction
+
+  #***********************************************
+  # Simulated prediction
+  #***********************************************
+
+  if (type == "simulated") {
+
+    if (progress) {
+      cat("\n")
+      pb <- utils::txtProgressBar(min = 0, max = M, style = 3)
+    }
+
+    if (missing(ci)) {
+      ci <- 0.95
+    } else {
+      if (!is.numeric(ci) || (ci < 0) || (ci > 1)) {
+        stop("ci argument has been misspecified.\n")
+      }
+    }
+    alpha <- (1 - ci) / 2
+
+    b.curr <- delta.prop <- b.hat$par
+    sigma.prop <- -solve(b.hat$hessian) * scale
+    accept <- 0
+    surv <- matrix(nrow = length(u), ncol = M)
+    for (m in 1:M) {
+      # Step 1: draw theta
+      theta.samp <- theta_draw(object)
+      # Step 2i: draw b from proposal distribution
+      b.prop <- b_draw(object, theta.samp, data.t, scale)
+      # Step 2ii: M-H acceptance
+      log.a1 <- logpb(b.prop, theta.samp, data.t) - logpb(b.curr, object$coefficients, data.t)
+      dens.curr <- mvtnorm::dmvt(x = b.curr,
+                                 delta = delta.prop,
+                                 sigma = sigma.prop,
+                                 df = 4,
+                                 log = TRUE)
+      dens.prop <- mvtnorm::dmvt(x = b.prop,
+                                 delta = delta.prop,
+                                 sigma = sigma.prop,
+                                 df = 4,
+                                 log = TRUE)
+      log.a2 <- dens.curr - dens.prop
+      log.a <- min(log.a1 - log.a2, 0)
+      log.u <- log(runif(1))
+      if (log.u <= log.a) {
+        b.curr <- b.prop
+        accept <- accept + 1
+      }
+      # Step 3: predicted survival
+      S.t <- S(b = b.curr, data = data.t, theta = theta.samp)
+      S.u <- vector(length = length(u))
+      for (i in 1:length(u)) {
+        data.u <- process_newdata(newdata = newdata,
+                                  object = object,
+                                  tobs = u[i],
+                                  newSurvData = newSurvData)
+        S.u[i] <- S(b = b.curr, data = data.u, theta = theta.samp)
+      }
+      surv[, m] <- S.u / S.t
+      if (progress) {
+        utils::setTxtProgressBar(pb, m)
+      }
+    }
+
+    if (progress) {
+      close(pb)
+    }
+
+    surv.mean <- apply(surv, 1, mean)
+    surv.med <- apply(surv, 1, median)
+    surv.low <- apply(surv, 1, quantile, prob = alpha)
+    surv.upp <- apply(surv, 1, quantile, prob = 1 - alpha)
+    pred <- data.frame("u" = u,
+                       "mean" = surv.mean,
+                       "median" = surv.med,
+                       "lower" = surv.low,
+                       "upper" = surv.upp)
+
+  } # end simulated prediction
 
   out <- list("pred" = pred,
               "fit" = object,
               "newdata" = newdata,
               "newSurvData" = newSurvData,
               "u" = u,
-              "data.t" = data.t)
+              "data.t" = data.t,
+              "type" = type)
+
+  if (type == "simulated") {
+    out$M <- ifelse(type == "simulated", M, NULL)
+    out$accept <- ifelse(type == "simulated", accept / M, NULL)
+  }
 
   class(out) <- "dynSurv"
   return(out)