+++ R/logpb.R
@@ -58,22 +58,6 @@ logpb <- function(b, theta, data) {
 }
 
 
-#' @keywords internal
-b.mode <- function(theta, data) {
-
-  out <- optim(par = rep(0, sum(data$r)),
-               fn = logpb,
-               theta = theta,
-               data = data,
-               control = list(fnscale = -1),
-               method = "BFGS",
-               hessian = TRUE)
-
-  return(out)
-
-}
-
-