+++ R/print.dynSurv.R
@@ -7,6 +7,11 @@ print.dynSurv <- function(x, digits = max(4, getOption("digits") - 4), ...) {
   }
 
   print(round(x$pred, digits = digits))
+
+  if (x$type == "simulated") {
+    cat(paste0("\nM-H acceptance rate: ", round(100 * x$accept, 1), "%"))
+  }
+
   cat("\n")
   invisible(x)